refractor academic form
