+++ camply-web/src/components/academic-form/components/form-inputs/LargeLabel.tsx
@@ -1,0 +1,15 @@
+import { cn } from '../../../../lib/utils';
+
+interface LargeLabelProps extends React.LabelHTMLAttributes<HTMLLabelElement> {}
+
+export const LargeLabel = ({ children, ...props }: LargeLabelProps) => (
+  <label 
+    {...props}
+    className={cn(
+      "text-xl md:text-2xl font-semibold text-white drop-shadow-lg mb-3 block",
+      props.className
+    )}
+  >
+    {children}
+  </label>
+);
