+++ camply-web/src/components/academic-form/components/form-inputs/index.ts
@@ -1,0 +1,3 @@
+export { LargeInput } from './LargeInput';
+export { LargeSelect } from './LargeSelect';
+export { LargeLabel } from './LargeLabel';
