+++ camply-web/src/components/academic-form/components/steps/index.ts
@@ -1,0 +1,6 @@
+export { NameStep } from './NameStep';
+export { PhoneStep } from './PhoneStep';
+export { CollegeStep } from './CollegeStep';
+export { DepartmentStep } from './DepartmentStep';
+export { RollNumberStep } from './RollNumberStep';
+export { YearsStep } from './YearsStep';
