+++ camply-web/src/components/academic-form/index.ts
@@ -1,0 +1,2 @@
+export { default } from './AcademicDetailsForm';
+export type { AcademicDetailsFormProps } from './types';
